feat(JSAM-74): restructure sign up files
