+++ JSA-Mawsitsit-Frontend/src/components/SignUp/SignUp.js
@@ -1,0 +1,16 @@
+// External Dependencies
+import React from 'react';
+// import PropTypes from 'prop-types';
+import { View } from 'react-native';
+
+// Internal Dependencies
+import SignUpForm from './SignUpForm';
+
+// Component Definitions
+const SignUp = () => (
+  <View>
+    <SignUpForm />
+  </View>
+);
+
+export default SignUp;
