+++ JSA-Mawsitsit-Frontend/src/components/SignUp/SignUpOld.js
@@ -1,0 +1,243 @@
+// External Dependencies
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import {
+  Text,
+  TextInput,
+  View,
+  ImageBackground,
+} from 'react-native';
+import { Button } from 'native-base';
+
+// Internal Dependencies
+import { URL } from 'react-native-dotenv';
+import Footer from './Footer';
+import {
+  emailValidate,
+  phoneNumberValidate,
+  passwordValidate,
+  passwordCheck,
+} from '../../../validation/validation';
+import PasswordStrength from './PasswordStrength';
+import RegisterStyle from './styles/SignUpStyle';
+
+// Local Variables
+const {
+  rootStyle,
+  warningTextColor,
+  containerStyle,
+  textStyle,
+  textInputStyle,
+  warningBorderColor,
+  buttonTextStyle,
+  buttonContainerStyle,
+  buttonStyle,
+} = RegisterStyle;
+
+const initInputText = {
+  email: '',
+  phoneNumber: '',
+  password: '',
+  passwordConfirm: '',
+};
+const initInputError = {
+  emailError: '',
+  phoneNumberError: '',
+  passwordMessage: '',
+  checkMessage: '',
+};
+
+// Component Definitions
+const RegisterScreen = (props) => {
+  const { navigation } = props;
+  const [inputText, setInputText] = useState(initInputText);
+  const [inputError, setInputError] = useState(initInputError);
+  const validated = (!Object.values(inputText).includes(''))
+    && (!inputError.emailError && !inputError.phoneNumberError && !inputError.checkMessage);
+  const handleChange = (name, value) => {
+    setInputText({
+      ...inputText,
+      [name]: value,
+    });
+  };
+
+  const postResult = (object) => {
+    fetch(`${URL}/register`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(object),
+    });
+  };
+
+  const handleSubmit = () => {
+    const register = {
+      email: inputText.email,
+      phone_number: inputText.phoneNumber,
+      password: inputText.password,
+    };
+    setInputText(initInputText);
+    setInputError(initInputError);
+    postResult(register);
+  };
+
+  return (
+    <View style={rootStyle}>
+      <ImageBackground
+        source={{ uri: 'https://i.pinimg.com/564x/a7/3f/b6/a73fb617b48b1e37f39f380f3ba448f3.jpg' }}
+        style={{ width: '100%', height: '100%' }}
+        imageStyle={{ opacity: 0.7 }}
+      >
+        <View style={containerStyle}>
+          <Text style={inputError.emailError ? { ...textStyle, ...warningTextColor } : textStyle}>
+            E-mail
+          </Text>
+          <TextInput
+            style={inputError.emailError
+              ? { ...textInputStyle, ...warningBorderColor } : textInputStyle}
+            value={inputText.email}
+            textContentType="emailAddress"
+            autoCapitalize="none"
+            placeholder="Email"
+            onChangeText={(value) => handleChange('email', value)}
+            onBlur={() => {
+              const emailError = emailValidate(inputText.email);
+              setInputError({
+                ...inputError,
+                emailError,
+              });
+            }}
+          />
+          <View>
+            {Boolean(inputError.emailError)
+              && <Text style={{ ...textStyle, ...warningTextColor }}>{inputError.emailError}</Text>}
+          </View>
+          <Text
+            style={inputError.phoneNumberError ? { ...textStyle, ...warningTextColor } : textStyle}
+          >
+            Phone number
+          </Text>
+          <TextInput
+            style={inputError.phoneNumberError
+              ? { ...textInputStyle, ...warningBorderColor } : textInputStyle}
+            value={inputText.phoneNumber}
+            textContentType="telephoneNumber"
+            autoCapitalize="none"
+            placeholder="phone number"
+            onChangeText={(value) => handleChange('phoneNumber', value)}
+            onBlur={() => {
+              const phoneNumberError = phoneNumberValidate(inputText.phoneNumber);
+              setInputError({
+                ...inputError,
+                phoneNumberError,
+              });
+            }}
+          />
+          <View>
+            {Boolean(inputError.phoneNumberError) && (
+              <Text style={{ ...textStyle, ...warningTextColor }}>
+                {inputError.phoneNumberError}
+              </Text>
+            )}
+          </View>
+          <Text
+            style={inputError.passwordMessage ? { ...textStyle, ...warningTextColor } : textStyle}
+          >
+            Password
+          </Text>
+          <TextInput
+            style={inputError.passwordMessage
+              ? { ...textInputStyle, ...warningBorderColor } : textInputStyle}
+            value={inputText.password}
+            secureTextEntry
+            textContentType="password"
+            autoCapitalize="none"
+            placeholder="password"
+            onChangeText={(value) => {
+              handleChange('password', value);
+              const passwordMessage = passwordValidate(value);
+              setInputError({
+                ...inputError,
+                passwordMessage,
+              });
+            }}
+            onBlur={() => {
+              const passwordMessage = passwordValidate(inputText.password);
+              setInputError({
+                ...inputError,
+                passwordMessage,
+              });
+            }}
+          />
+
+          <View>
+            {Boolean(inputError.passwordMessage)
+              && (<PasswordStrength passwordMessage={inputError.passwordMessage} />)}
+          </View>
+
+          <Text style={inputError.checkMessage ? { ...textStyle, ...warningTextColor } : textStyle}>
+            Password confirmation
+          </Text>
+          <TextInput
+            style={inputError.checkMessage
+              ? { ...textInputStyle, ...warningBorderColor } : textInputStyle}
+            value={inputText.passwordConfirm}
+            secureTextEntry
+            textContentType="password"
+            autoCapitalize="none"
+            placeholder="password"
+            onChangeText={(value) => {
+              handleChange('passwordConfirm', value);
+              const checkMessage = passwordCheck(inputText.password, value);
+              setInputError({
+                ...inputError,
+                checkMessage,
+              });
+            }}
+            onBlur={() => {
+              const checkMessage = passwordCheck(inputText.password, inputText.passwordConfirm);
+              setInputError({
+                ...inputError,
+                checkMessage,
+              });
+            }}
+          />
+          <View>
+            {Boolean(inputError.checkMessage) && (
+              <Text style={{ ...textStyle, ...warningTextColor }}>
+                {inputError.checkMessage}
+              </Text>
+            )}
+          </View>
+        </View>
+
+        <View style={buttonContainerStyle}>
+          <Button
+            style={buttonStyle}
+            onPress={handleSubmit}
+            disabled={!validated}
+          >
+            <Text style={buttonTextStyle}>Sign Up</Text>
+          </Button>
+          <Footer navigation={navigation} />
+        </View>
+      </ImageBackground>
+
+    </View>
+  );
+};
+
+// Navigation Configuration
+RegisterScreen.navigationOptions = {
+  headerShown: false,
+};
+
+RegisterScreen.propTypes = {
+  navigation: PropTypes.shape({
+    navigate: PropTypes.func,
+  }),
+};
+RegisterScreen.defaultProps = {
+  navigation: {},
+};
+
+export default RegisterScreen;
